Splits toArray to its own module
