+++ src/lib/rovingTabIndex.test.js
@@ -1,10 +1,5 @@
 /* eslint-disable max-len */
-import {
-  rovingTabIndex,
-  tabIndexAllow,
-  tabIndexDeny,
-  toArray,
-} from './rovingTabIndex';
+import { rovingTabIndex, tabIndexAllow, tabIndexDeny } from './rovingTabIndex';
 
 // Set up our document body
 document.body.innerHTML = `
@@ -16,7 +11,6 @@ document.body.innerHTML = `
 `;
 
 const elements = document.querySelectorAll('span');
-const div = document.querySelector('div');
 
 describe('Correctly updates tabindex attribute.', () => {
   it('Allow on a single element', () => {
@@ -42,11 +36,4 @@ describe('Correctly updates tabindex attribute.', () => {
     expect(elements[1].getAttribute('tabindex')).toEqual('-1');
     expect(elements[2].getAttribute('tabindex')).toEqual('-1');
   });
-});
-
-it('Should convert an HTMLElement and NodeList to an Array', () => {
-  expect(toArray(elements)).toEqual(expect.arrayContaining([elements[0], elements[1], elements[2]]));
-  expect(toArray(elements[1])).toEqual(expect.arrayContaining([elements[1]]));
-  expect(toArray(elements[1].classList)).toEqual(expect.arrayContaining([]));
-  expect(toArray(div.children)).toEqual(expect.arrayContaining([elements[0], elements[1], elements[2]]));
 });
